+++ api/job_events_test.go
@@ -54,20 +54,19 @@ func TestJobEventsSubscribe(t *testing.T) {
 
 			var wg sync.WaitGroup
 			expectedTags := []string{"tag1", "tag2"}
-			ch, _ := events.AddSubscriber(context.Background(), events.FilterParams{ClientIDs: []string{clientID}}, expectedTags)
+			subscriber := events.AddSubscriber(t.Context(), time.Minute, events.FilterParams{ClientIDs: []string{clientID}}, expectedTags)
 			wg.Add(1)
 			go func() {
 				defer wg.Done()
 
 				// wait for job created event
-				ev := <-ch
-				payload := ev.Args[0].(*events.JobEvent)
+				payload := <-subscriber.Events
 				assert.Equal(t, events.ActionCreate, payload.Action)
 				assert.Equal(t, expectedTags, payload.Tags)
 				jobID.Store(&payload.Job.ID)
 
 				// wait for event created by our status.Update below
-				<-ch
+				<-subscriber.Events
 				// now our GET request should have received the response as well,
 				// add some extra time to be safe
 				time.Sleep(100 * time.Millisecond)
@@ -95,10 +94,17 @@ func TestJobEventsSubscribe(t *testing.T) {
 			}
 
 			req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("/api/wfx/v1/jobs/events?ids=%s", *jobID.Load()), nil)
-			rec := sse.NewMockResponseRecorder()
+			rec := sse.NewMockResponseRecorder(t)
 			handler.ServeHTTP(rec, req)
 
-			response := <-rec.ChResponse
+			var response string
+			for {
+				response = rec.Response()
+				if strings.Contains(response, "id: 1") {
+					break
+				}
+			}
+
 			assert.Contains(t, response, "HTTP/1.1 200")
 			assert.Contains(t, response, "Content-Type: text/event-stream")
 
@@ -107,7 +113,7 @@ func TestJobEventsSubscribe(t *testing.T) {
 			for _, line := range lines {
 				t.Logf(">> %s", line)
 			}
-			assert.Len(t, lines, 7)
+			assert.Len(t, lines, 8)
 
 			lines = strings.Split(lines[len(lines)-1], "\n")