+++ internal/handler/job/events/events.go
@@ -10,15 +10,55 @@ package events
 
 import (
 	"context"
+	"sync"
+	"time"
 
 	"github.com/go-openapi/strfmt"
 	"github.com/google/uuid"
-	"github.com/olebedev/emitter"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/siemens/wfx/generated/api"
 	"github.com/siemens/wfx/middleware/logging"
 )
+
+type Subscriber struct {
+	Events  <-chan JobEvent // channel which receives the events for the subscriber
+	Backlog *Backlog
+
+	id            string        // unique identifier
+	ch            chan JobEvent // internal RW of Events
+	lastBlock     time.Time     // time of the last failed attempt to send a message
+	graceInterval time.Duration // duration after which non-responsive subscribers are dropped
+
+	jobIDSet    map[string]any // job filter
+	clientIDSet map[string]any // clientID filter
+	workflowSet map[string]any // workflow filter
+	tags        []string       // tags to apply
+}
+
+type Backlog struct {
+	data []JobEvent
+	mu   sync.Mutex
+}
+
+// Enq adds a new JobEvent to the backlog.
+func (b *Backlog) Enq(event JobEvent) {
+	b.mu.Lock()
+	b.data = append(b.data, event)
+	b.mu.Unlock()
+}
+
+// Deq removes and returns the oldest (FIFO) JobEvent from the backlog.
+func (b *Backlog) Deq() (*JobEvent, bool) {
+	b.mu.Lock()
+	defer b.mu.Unlock()
+	if len(b.data) == 0 {
+		return nil, false
+	}
+	var ev JobEvent
+	ev, b.data = b.data[0], b.data[1:]
+	return &ev, true
+}
 
 type JobEvent struct {
 	// Ctime is the time when the event was created
@@ -45,16 +85,16 @@ const (
 	ActionUpdateDefinition Action = "UPDATE_DEFINITION"
 )
 
-// NOTE: The number 32 for capacity is arbitrary; suggestions for a different value are welcome, particularly if supported by evidence.
-var e *emitter.Emitter = emitter.New(32)
+var (
+	subscribers   = make([]*Subscriber, 0)
+	muSubscribers sync.RWMutex // mutex used for locking subscribers
+)
 
 // AddSubscriber adds a new subscriber to receive job events filtered based on the provided filterParams.
-func AddSubscriber(ctx context.Context, filter FilterParams, tags []string) (<-chan emitter.Event, error) {
+func AddSubscriber(ctx context.Context, graceInterval time.Duration, filter FilterParams, tags []string) *Subscriber {
 	log := logging.LoggerFromCtx(ctx)
-
 	// for logging purposes
 	subscriberID := uuid.New().String()
-
 	log.Info().
 		Str("subscriberID", subscriberID).
 		Dict("filterParams", zerolog.Dict().
@@ -64,48 +104,9 @@ func AddSubscriber(ctx context.Context, filter FilterParams, tags []string) (<-c
 		Strs("tags", tags).
 		Msg("Adding new subscriber for job events")
 
-	if filter := filter.createEventFilter(subscriberID, tags); filter != nil {
-		return e.On("*", filter), nil
-	}
-	return e.On("*"), nil
-}
+	ch := make(chan JobEvent, 1)
 
-// ShutdownSubscribers disconnects all subscribers.
-func ShutdownSubscribers() {
-	count := len(e.Topics())
-	for _, topic := range e.Topics() {
-		log.Debug().Str("topic", topic).Msg("Closing subscribers")
-		e.Off(topic)
-	}
-	log.Info().Int("count", count).Msg("Subscriber shutdown complete")
-}
-
-// SubscriberCount counts the total number of subscribers across all topics.
-func SubscriberCount() int {
-	count := 0
-	for _, topic := range e.Topics() {
-		count += len(e.Listeners(topic))
-	}
-	return count
-}
-
-// PublishEvent publishes a new event. It returns a channel which can be used
-// to wait for the delivery of the event to all listeners.
-func PublishEvent(ctx context.Context, event *JobEvent) chan struct{} {
-	log := logging.LoggerFromCtx(ctx)
-	log.Debug().
-		Str("jobID", event.Job.ID).
-		Str("action", string(event.Action)).Msg("Publishing event")
-	return e.Emit(event.Job.ID, event)
-}
-
-func (filter FilterParams) createEventFilter(subscriberID string, tags []string) func(*emitter.Event) {
-	// special case: no filters means "catch-all"
-	if len(filter.JobIDs) == 0 && len(filter.ClientIDs) == 0 && len(filter.Workflows) == 0 {
-		return nil
-	}
-
-	// build hash sets for faster lookup
+	// build hash maps for faster lookup
 	jobIDSet := make(map[string]any, len(filter.JobIDs))
 	for _, s := range filter.JobIDs {
 		jobIDSet[s] = nil
@@ -118,27 +119,125 @@ func (filter FilterParams) createEventFilter(subscriberID string, tags []string)
 	for _, s := range filter.Workflows {
 		workflowSet[s] = nil
 	}
-	return func(ev *emitter.Event) {
-		event := ev.Args[0].(*JobEvent)
-		job := event.Job
-		// check if we shall notify the client about the event
-		_, interested := jobIDSet[job.ID]
-		if !interested {
-			_, interested = clientIDSet[job.ClientID]
+
+	subscriber := &Subscriber{
+		id:            subscriberID,
+		Events:        ch,
+		ch:            ch,
+		Backlog:       &Backlog{data: make([]JobEvent, 0)},
+		lastBlock:     time.Time{},
+		graceInterval: graceInterval,
+		jobIDSet:      jobIDSet,
+		clientIDSet:   clientIDSet,
+		workflowSet:   workflowSet,
+		tags:          tags,
+	}
+
+	muSubscribers.Lock()
+	subscribers = append(subscribers, subscriber)
+	muSubscribers.Unlock()
+
+	return subscriber
+}
+
+// ShutdownSubscribers disconnects all subscribers.
+func ShutdownSubscribers() {
+	muSubscribers.Lock()
+	oldSubscribers := subscribers
+	subscribers = make([]*Subscriber, 0)
+	muSubscribers.Unlock()
+
+	for _, subscriber := range oldSubscribers {
+		close(subscriber.ch)
+	}
+	log.Info().Msg("Subscriber shutdown complete")
+}
+
+func RemoveSubscriber(subscriber *Subscriber) {
+	log.Debug().Str("id", subscriber.id).Msg("Removing subscriber")
+
+	muSubscribers.Lock()
+	defer muSubscribers.Unlock()
+
+	oldSubscribers := subscribers
+	subscribers = make([]*Subscriber, 0, len(subscribers))
+	for _, sub := range oldSubscribers {
+		if sub == subscriber {
+			continue
 		}
-		if !interested && job.Workflow != nil {
-			_, interested = workflowSet[job.Workflow.Name]
+		subscribers = append(subscribers, sub)
+	}
+	close(subscriber.ch)
+}
+
+// SubscriberCount counts the total number of subscribers across all topics.
+func SubscriberCount() int {
+	muSubscribers.RLock()
+	defer muSubscribers.RUnlock()
+	return len(subscribers)
+}
+
+// PublishEvent publishes a new event. This is a synchronous operation.
+func PublishEvent(ctx context.Context, event JobEvent) {
+	log := logging.LoggerFromCtx(ctx).With().Str("jobID", event.Job.ID).Str("action", string(event.Action)).Logger()
+	log.Debug().Msg("Publishing event to subscribers")
+
+	muSubscribers.Lock()
+	defer muSubscribers.Unlock()
+
+	// the subscribers that are still active and we'll keep
+	newSubscribers := make([]*Subscriber, 0, len(subscribers))
+
+	for _, sub := range subscribers {
+		ctxLog := log.With().Str("id", sub.id).Logger()
+
+		// check if we shall notify the subscriber about the event
+		// special case: no filters means "catch-all"
+		isCatchAll := len(sub.jobIDSet) == 0 && len(sub.clientIDSet) == 0 && len(sub.workflowSet) == 0
+		interested := isCatchAll ||
+			mapContains(sub.jobIDSet, event.Job.ID) ||
+			mapContains(sub.clientIDSet, event.Job.ClientID) ||
+			(event.Job.Workflow != nil && mapContains(sub.workflowSet, event.Job.Workflow.Name))
+		if !interested {
+			// keep subscriber, potentially still alive
+			newSubscribers = append(newSubscribers, sub)
+			ctxLog.Debug().Msg("Subscriber not interested, skipping event notification")
+			continue
 		}
 
-		if interested {
-			// apply tags
-			log.Debug().Strs("tags", tags).Msg("Applying tags to event notification")
-			event.Tags = tags
+		// apply tags
+		event.Tags = sub.tags
+
+		// try to send event
+		ctxLog.Debug().Msg("Sending event to subscriber")
+		keepSubscriber := true
+		select {
+		case sub.ch <- event:
+			ctxLog.Info().Msg("Sent event to subscriber")
+			// reset lastBlock
+			sub.lastBlock = time.Time{}
+		default: // unable to send event (channel full)
+			ctxLog.Info().Msg("Unable to send event, channel is full")
+			// add to subscriber's backlog
+			ctxLog.Debug().Msg("Adding event to subscriber backlog")
+			sub.Backlog.Enq(event)
+			// check if we should drop the subscriber
+			keepSubscriber = sub.lastBlock.IsZero() || time.Since(sub.lastBlock) <= sub.graceInterval
+			sub.lastBlock = time.Now()
+		}
+		if keepSubscriber {
+			ctxLog.Debug().Msg("Keeping subscriber")
+			newSubscribers = append(newSubscribers, sub)
 		} else {
-			log.Debug().
-				Str("subscriberID", subscriberID).
-				Str("jobID", job.ID).Msg("Skipping event notification")
-			emitter.Void(ev)
+			ctxLog.Info().Msg("Dropping inactive subscriber")
+			close(sub.ch)
 		}
 	}
+
+	subscribers = newSubscribers
+}
+
+func mapContains(set map[string]any, key string) (found bool) {
+	_, found = set[key]
+	return
 }